+++ src/bind.c
@@ -64,7 +64,12 @@ static SEXP vec_rbind(SEXP xs, SEXP ptype, SEXP names_to, struct name_repair_opt
 
   // Find individual input sizes and total size of output
   R_len_t nrow = 0;
+
   bool has_rownames = false;
+  if (names_to == R_NilValue && r_names(xs) != R_NilValue) {
+    // Names of inputs become row names when `names_to` isn't supplied
+    has_rownames = true;
+  }
 
   SEXP ns_placeholder = PROTECT_N(Rf_allocVector(INTSXP, n), &nprot);
   int* ns = INTEGER(ns_placeholder);
@@ -89,14 +94,25 @@ static SEXP vec_rbind(SEXP xs, SEXP ptype, SEXP names_to, struct name_repair_opt
     rownames = PROTECT_N(Rf_allocVector(STRSXP, nrow), &nprot);
   }
 
+  SEXP nms = PROTECT_N(r_names(xs), &nprot);
+  bool has_names = nms != R_NilValue;
+  bool has_names_to = names_to != R_NilValue;
+
+  const SEXP* nms_p = NULL;
+  if (has_names) {
+    nms_p = STRING_PTR_RO(nms);
+  }
+
   SEXP names_to_col = R_NilValue;
   SEXPTYPE names_to_type = 99;
   void* names_to_p = NULL;
   const void* index_p = NULL;
 
-  if (names_to != R_NilValue) {
-    SEXP index = PROTECT_N(r_names(xs), &nprot);
-    if (index == R_NilValue) {
+  if (has_names_to) {
+    SEXP index = R_NilValue;
+    if (has_names) {
+      index = nms;
+    } else {
       index = PROTECT_N(Rf_allocVector(INTSXP, n), &nprot);
       r_int_fill_seq(index, 1, n);
     }
@@ -123,13 +139,27 @@ static SEXP vec_rbind(SEXP xs, SEXP ptype, SEXP names_to, struct name_repair_opt
 
     if (has_rownames) {
       SEXP rn = df_rownames(x);
+
+      if (has_names && nms_p[i] != strings_empty && !has_names_to) {
+        if (rownames_type(rn) == ROWNAMES_IDENTIFIERS) {
+          rn = r_chr_paste_prefix(rn, CHAR(nms_p[i]), "...");
+        } else if (size > 1) {
+          rn = r_seq_chr(CHAR(nms_p[i]), size);
+        } else {
+          rn = r_str_as_character(nms_p[i]);
+        }
+      }
+      PROTECT(rn);
+
       if (rownames_type(rn) == ROWNAMES_IDENTIFIERS) {
         chr_assign(rownames, idx, rn, false);
       }
+
+      UNPROTECT(1);
     }
 
     // Assign current name to group vector, if supplied
-    if (names_to != R_NilValue) {
+    if (has_names_to) {
       r_vec_fill(names_to_type, names_to_p, index_p, i, size);
       r_vec_ptr_inc(names_to_type, &names_to_p, size);
     }
@@ -143,7 +173,7 @@ static SEXP vec_rbind(SEXP xs, SEXP ptype, SEXP names_to, struct name_repair_opt
     Rf_setAttrib(out, R_RowNamesSymbol, rownames);
     UNPROTECT(1);
   }
-  if (names_to != R_NilValue) {
+  if (has_names_to) {
     out = df_poke_at(out, names_to, names_to_col);
   }