+++ tests/testthat/test-bind.R
@@ -200,6 +200,34 @@ test_that("row names are preserved by vec_rbind()", {
   expect_identical(vec_rbind(df1, df2), out)
 })
 
+test_that("can assign row names in vec_rbind()", {
+  df1 <- mtcars[1:3, ]
+  df2 <- mtcars[4:5, ]
+
+  # Combination
+  out <- vec_rbind(foo = df1, df2)
+  exp <- mtcars[1:5, ]
+  row.names(exp) <- c(paste0("foo...", row.names(df1)), row.names(df2))
+  expect_identical(out, exp)
+
+  out <- vec_rbind(foo = df1, df2, .names_to = "id")
+  exp <- mtcars[1:5, ]
+  exp$id <- c(rep("foo", 3), rep("", 2))
+  expect_identical(out, exp)
+
+  # Sequence
+  out <- vec_rbind(foo = unrownames(df1), df2, bar = unrownames(mtcars[6, ]))
+  exp <- mtcars[1:6, ]
+  row.names(exp) <- c(paste0("foo", 1:3), row.names(df2), "bar")
+  expect_identical(out, exp)
+
+  out <- vec_rbind(foo = unrownames(df1), df2, bar = unrownames(mtcars[6, ]), .names_to = "id")
+  exp <- mtcars[1:6, ]
+  exp$id <- c(rep("foo", 3), rep("", 2), "bar")
+  row.names(exp) <- c(paste0("...", 1:3), row.names(df2), "...6")
+  expect_identical(out, exp)
+})
+
 
 # cols --------------------------------------------------------------------
 
@@ -373,20 +401,19 @@ test_that("vec_cbind() fails with arrays of dimensionality > 3", {
 })
 
 test_that("vec_rbind() consistently handles unnamed outputs", {
-  skip("FIXME")
-  # These are a little weird but unclear we can do better
+  # Name repair of columns is a little weird but unclear we can do better
   expect_identical(
     vec_rbind(1, 2),
     data.frame(...1 = c(1, 2))
   )
   expect_identical(
     vec_rbind(1, 2, ...10 = 3),
-    data.frame(...1 = c(1, 2, 3))
+    data.frame(...1 = c(1, 2, 3), row.names = c("...1", "...2", "...3"))
   )
 
   expect_identical(
     vec_rbind(a = 1, b = 2),
-    data.frame(a = c(1, NA), b = c(NA, 2))
+    data.frame(...1 = c(1, 2), row.names = c("a", "b"))
   )
   expect_identical(
     vec_rbind(c(a = 1), c(b = 2)),