+++ NEWS.md
@@ -1,7 +1,9 @@
 
 # vctrs (development version)
 
-* `vec_rbind()` now binds row names.
+* `vec_rbind()` now binds row names. When named inputs are supplied
+  and `names_to` is `NULL`, the names define row names. If `names_to`
+  is supplied, they are assigned in the column name as before.
 
 * The `c()` method for `vctrs_vctr` now throws an error when
   `recursive` or `use.names` is supplied (#791).